# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: API per il servizio di messaggistica "WASAText" del corso WASA.
  version: "1.0.0"
servers:
  - url: /
...
security: # Regole di sicurezza globali
  - bearerAuth: [] # Definito in components/securitySchemes/bearerAuth
paths:
...
  /settings/username:
    put:
      tags: ["settings"] 
      summary: Set my username
      description: Aggiorna il nome utente dell'utente autenticato.
      operationId: setMyUsername
      requestBody:
        description: Il nuovo username desiderato.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetMyUsernameRequest'
      responses:
        '200':
          description: Username aggiornato con successo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          $ref: '#/components/responses/Conflict'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /settings/photo:
    put:
      tags: ["settings"] 
      summary: Set my profile photo
      description: Aggiorna l'URL della foto profilo dell'utente autenticato.
      operationId: setMyPhoto
      requestBody:
        description: Il nuovo URL della foto profilo.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPhotoRequest'
      responses:
        '200':
          description: Foto aggiornata con successo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users: 
    get:
      tags: ["users"] 
      summary: Search users
      description: Cerca utenti in base a una parte del loro username.
      operationId: searchUsers 
      parameters:
        - $ref: '#/components/parameters/SearchUsername'
      responses:
        '200':
          description: Lista di utenti che corrispondono alla ricerca. Può essere vuota.
          content:
            application/json:
              schema:
                type: array
                description: Lista degli utenti trovati.
                items:
                  $ref: '#/components/schemas/User'
                minItems: 0
                maxItems: 100
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /conversations/{convId}/messages:
    post:
      tags: ["messages"]
      summary: Send a new message
      description: Invia un nuovo messaggio (testo o URL foto) in una conversazione specificata, a cui l'utente autenticato appartiene. Può opzionalmente essere una risposta a un messaggio esistente.
      operationId: sendMessage
      parameters:
        - $ref: '#/components/parameters/ConvID'
      requestBody:
        description: Contenuto del messaggio (testo o URL foto) e ID opzionale del messaggio a cui rispondere.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest' 
      responses:
        '201':
          description: Messaggio inviato con successo. Restituisce il messaggio creato.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /conversations/{convId}/forwarded-messages:
    post:
      tags: ["messages"]
      summary: Forward a message
      description: Inoltra un messaggio esistente (identificato dal suo ID) in una conversazione specificata, a cui l'utente autenticato appartiene.
      operationId: forwardMessage
      parameters:
        - $ref: '#/components/parameters/ConvID'
      requestBody:
        description: L'ID del messaggio originale da inoltrare.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardMessageRequest'
      responses:
        '201':
          description: Messaggio inoltrato con successo. Restituisce il nuovo messaggio creato (l'inoltro).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /messages/{msgId}:
    delete:
      tags: ["messages"] 
      summary: Delete a message
      description: Cancella un messaggio inviato dall'utente autenticato. 
      operationId: deleteMessage
      parameters:
        - $ref: '#/components/parameters/MsgID'
      responses:
        '204':
          description: Messaggio cancellato con successo. Nessun contenuto restituito.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /messages/{msgId}/reactions:
    post:
      tags: ["messages"]
      summary: React to a message
      description: Aggiunge una reazione (emoji) a un messaggio specificato, a nome dell'utente autenticato. Se l'utente ha già reagito con la stessa emoji, non succede nulla.
      operationId: commentMessage
      parameters:
        - $ref: '#/components/parameters/MsgID'
      requestBody:
        description: L'emoji della reazione.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentMessageRequest' 
      responses:
        '201':
          description: Reazione aggiunta/aggiornata con successo. Restituisce la reazione creata/aggiornata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reaction'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /messages/{msgId}/reactions/{reactionId}:
    delete:
      tags: ["messages"]
      summary: Remove a reaction
      description: Rimuove una reazione (identificata dal suo ID) da un messaggio specificato. L'operazione è permessa solo all'utente che ha creato quella reazione.
      operationId: uncommentMessage
      parameters:
        - $ref: '#/components/parameters/MsgID'
        - $ref: '#/components/parameters/ReactionID'
      responses:
        '204':
          description: Reazione rimossa con successo. Nessun contenuto restituito.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "string"
      description: "L'identificativo utente (ottenuto dal login) da usare come Bearer token nell'header `Authorization: Bearer <identifier>`."
  parameters:
    ConvID:
      name: convId
      in: path
      required: true
      description: ID univoco della conversazione.
      schema:
        $ref: '#/components/schemas/InternalID' 
    MsgID:
      name: msgId
      in: path
      required: true
      description: ID univoco del messaggio.
      schema:
        $ref: '#/components/schemas/InternalID'
    ReactionID:
      name: reactionId
      in: path
      required: true
      description: ID univoco della reazione.
      schema:
        $ref: '#/components/schemas/InternalID'
    SearchUsername:
      name: username
      in: query
      required: true
      description: Stringa da cercare nel nome utente (case-insensitive).
      schema:
        $ref: '#/components/schemas/Username'
  responses:
    BadRequest: # 400
      description: Formato dei dati non valido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized: # 401
      description: Autenticazione richiesta o token non valido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden: # 403
      description: Non hai i permessi per eseguire questa azione.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound: # 404
      description: La risorsa richiesta non è stata trovata.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict: # 409
      description: La risorsa esiste già o è in conflitto.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError: # 500
      description: Errore interno del server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UserID:
      description: ID univoco in formato UUID (usato per utenti e bearer token).
      type: string
      pattern: '^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$'
      minLength: 36
      maxLength: 36
      example: "abcdef01-2345-6789-abcd-ef0123456789"
    InternalID:
      description: ID univoco generato dal sistema (per conversazioni, messaggi, reazioni).
      type: string
      pattern: '^.+$'
      minLength: 1
      maxLength: 50
      example: "msg-789"
    Username:
      description: Nome univoco scelto dall'utente.
      type: string
      pattern: '^.+$'
      minLength: 3
      maxLength: 16
      example: "Maria"
    PhotoURL:
      description: URL di un'immagine (foto profilo o messaggio).
      type: string
      format: uri
      pattern: '^.+$'
      minLength: 1
      maxLength: 2048
      example: "https://example.com/photos/image.jpg"
    Emoji:
      description: Un singolo carattere emoji per una reazione.
      type: string
      pattern: '^.+$'
      minLength: 1
      maxLength: 4
      example: "👍"
    Timestamp:
      description: Data e ora in formato ISO 8601 UTC.
      type: string
      format: date-time
      pattern: '^.+$'
      example: "2024-10-20T10:30:00Z"
...
    MessageContent:
      description: Contenuto testuale di un messaggio.
      type: string
      pattern: '^.+$'
      minLength: 1
      maxLength: 4000
      example: "Ciao!"


    # Strutture del Database (service/database/structs.go)
    User:
      description: Rappresenta un utente registrato nel sistema.
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UserID'
        username:
          $ref: '#/components/schemas/Username'
        photoUrl:
          $ref: '#/components/schemas/PhotoURL'
      required: ["id", "username"]
...
    Message:
      description: Rappresenta un singolo messaggio all'interno di una conversazione.
      type: object
      properties:
        id:
          $ref: '#/components/schemas/InternalID'
        sender:
          $ref: '#/components/schemas/User'
        content:
          $ref: '#/components/schemas/MessageContent'
        contentType:
          type: string
          description: Tipo di contenuto del messaggio.
          pattern: '^.+$'
          enum: ["text", "photo"]
          example: "text"
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        status:
          type: string
          description: Stato di consegna/lettura del messaggio (visibile solo per i messaggi inviati dall'utente autenticato). Non presente per i messaggi ricevuti.
          pattern: '^.+$'
          enum: ["sent", "received", "read"]
          example: "read"
        reactions:
          type: array
          description: Lista delle reazioni associate a questo messaggio.
          minItems: 0
          maxItems: 50
          items:
            $ref: '#/components/schemas/Reaction'
      required: ["id", "sender", "content", "contentType", "timestamp"]
    Reaction:
      description: Rappresenta una singola reazione (emoji) data da un utente a un messaggio.
      type: object
      properties:
        id:
          $ref: '#/components/schemas/InternalID'
        emoji:
          $ref: '#/components/schemas/Emoji'
        user:
          $ref: '#/components/schemas/User'
      required: ["id", "emoji", "user"]


    # Schemi di richiesta/risposta API (service/api/schemas.go)
    ErrorResponse:
      description: Schema generico per le risposte di errore.
      type: object
      properties:
        message:
          type: string
          description: Messaggio descrittivo dell'errore.
          minLength: 1
          maxLength: 256
          pattern: '^.+$'
          example: "Si è verificato un errore."
      required: ["message"]
...
    SetMyUsernameRequest:
      description: Dati inviati per aggiornare il nome utente.
      type: object
      properties:
        username:
          $ref: '#/components/schemas/Username'
      required: ["username"]
    SetPhotoRequest:
      description: Dati inviati per aggiornare l'URL della foto profilo dell'utente o del gruppo.
      type: object
      properties:
        photoUrl:
          $ref: '#/components/schemas/PhotoURL'
      required: ["photoUrl"]
...
    SendMessageRequest:
      description: Dati inviati per creare un nuovo messaggio. Deve contenere `text` o `photoUrl`.
      type: object
      properties:
        replyToMsgId:
          $ref: '#/components/schemas/InternalID'
      oneOf:
        - type: object
          description: Messaggio di testo.
          properties:
            text:
              $ref: '#/components/schemas/MessageContent'
          required: ["text"]
        - type: object
          description: Messaggio con foto.
          properties:
            photoUrl:
              $ref: '#/components/schemas/PhotoURL'
          required: ["photoUrl"]
    ForwardMessageRequest:
      description: Dati inviati per inoltrare un messaggio esistente.
      type: object
      properties:
        originalMessageId:
          $ref: '#/components/schemas/InternalID'
      required: ["originalMessageId"]
    CommentMessageRequest:
      description: Dati inviati per aggiungere una reazione a un messaggio.
      type: object
      properties:
        emoji:
          $ref: '#/components/schemas/Emoji'
      required: ["emoji"]